backend/requirements.txt
Flask
pytest
psycopg2-binary
flask-sqlalchemy
flask-cors
debugpy
python-dotenv
requests
Pyrebase4
firebase-admin